@@ -4,9 +4,11 @@
   "version": "0.8.14",
   "description": "Fully-featured, lightweight command-line test runner",
   "bin": {
-    "test-runner": "bin/cli.js"
+    "test-runner": "bin/cli.mjs"
   },
   "repository": "https://github.com/test-runner-js/test-runner",
+  "main": "./index.mjs",
+  "exports": "./index.mjs",
   "keywords": [
     "test-runner",
     "test",
@@ -22,15 +24,15 @@
   ],
   "license": "MIT",
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
   "files": [
-    "bin/*.js",
-    "lib/*.js",
-    "index.js"
+    "bin",
+    "lib",
+    "index.mjs"
   ],
   "scripts": {
-    "test": "node test/tests.js",
+    "test": "node test/tests.mjs",
     "docs": "jsdoc2md index.js > docs/API.md"
   },
   "dependencies": {
@@ -39,15 +41,16 @@
     "@test-runner/oneline-view": "^0.1.4",
     "command-line-args": "^5.1.1",
     "command-line-usage": "^6.1.1",
+    "current-module-paths": "^0.1.0",
     "file-set": "^4.0.2",
-    "load-module": "^3.0.0",
+    "load-module": "^4.0.0-0",
     "reduce-flatten": "^3.0.0",
-    "test-object-model": "^0.6.1",
-    "test-runner-core": "^0.8.6",
-    "walk-back": "^4.0.0"
+    "test-object-model": "^0.7.1",
+    "test-runner-core": "^0.9.0",
+    "walk-back": "^5.0.0"
   },
   "devDependencies": {
-    "jsdoc-to-markdown": "^7.0.0",
+    "jsdoc-to-markdown": "^7.0.1",
     "node-fetch": "^2.6.1"
   }
 }
