@@ -1,4 +1,4 @@
-const ansi = require('ansi-escape-sequences')
+import ansi from 'ansi-escape-sequences'
 
 class TreeView {
   constructor (tom) {
@@ -39,4 +39,4 @@ class TreeView {
   }
 }
 
-module.exports = TreeView
+export default TreeView
